Rename theme to styles so as not to confuse with victory-wide theme
