+++ demo/app.jsx
@@ -8,18 +8,18 @@ import useScroll from "scroll-behavior/lib/useStandardScroll";
 const appHistory = useScroll(useRouterHistory(createBrowserHistory))();
 // Routes
 import Index from "./index";
-import MultipleAxes from "./tutorials/multiple-axes";
-import CustomTheme from "./tutorials/custom-theme";
-import CustomDataComponent from "./tutorials/custom-data-component";
+import MultipleAxesDemo from "./examples/multiple-axes";
+import CustomStylesDemo from "./examples/custom-styles";
+import CustomDataComponentDemo from "./examples/custom-data-component";
 
 const content = document.getElementById("content");
 
 render((
   <Router history={appHistory}>
     <Route path="/" component={Index}>
-      <Route path="multiple-axes" component={MultipleAxes} />
-      <Route path="custom-theme" component={CustomTheme} />
-      <Route path="custom-data-component" component={CustomDataComponent} />
+      <Route path="multiple-axes" component={MultipleAxesDemo} />
+      <Route path="custom-styles" component={CustomStylesDemo} />
+      <Route path="custom-data-component" component={CustomDataComponentDemo} />
     </Route>
   </Router>
 ), content);
